Moves generic warning functions to more suitable module
